improve provide method for database object
